Fixed Boss Door not opening
Fixed by loading the door and its animation controller component after the scene loads by proximity instead of at the start of the game. Also had to change FindGameObjectWithTag for Find due to the former only working on the objects in the same scene even if more than one is loaded in the hierarchy.

Made sure the animation doesn't repeat every frame too, by addind a doorIsOpen boolean.